Rename `public` server to `forms` server
