+++ pkg/server/forms/handlers/database/delete.go
