+++ pkg/server/forms/handlers/folder/delete.go
