+++ pkg/server/forms/handlers/form/delete.go
