+++ pkg/server/forms/tests/main_test.go
@@ -6,7 +6,6 @@ import (
 	"github.com/nrc-no/core/pkg/client"
 	"github.com/nrc-no/core/pkg/rest"
 	"github.com/nrc-no/core/pkg/server/options"
-	"github.com/nrc-no/core/pkg/server/public"
 	"github.com/nrc-no/core/pkg/store"
 	"github.com/nrc-no/core/pkg/testutils"
 	"github.com/stretchr/testify/assert"
@@ -18,7 +17,7 @@ import (
 type Suite struct {
 	suite.Suite
 	cli    client.Client
-	srv    *public.Server
+	srv    *forms.Server
 	cancel context.CancelFunc
 	ctx    context.Context
 	done   func()
@@ -53,7 +52,7 @@ func (s *Suite) SetupSuite() {
 		s.T().FailNow()
 	}
 
-	s.srv, err = public.NewServer(public.Options{
+	s.srv, err = forms.NewServer(forms.Options{
 		ServerOptions: options.ServerOptions{
 			Host: "localhost",
 			Port: 0,
